Write additional test for missionsSlice component
